PO-2255 - fix merge issue

+++ src/integrationTest/java/uk/gov/hmcts/opal/service/GenericReportServiceTest.java
@@ -173,7 +173,7 @@ class GenericReportServiceTest extends AbstractIntegrationTest {
         }
 
         @Override
-        public int getNumberOfRecords() {
+        public long getNumberOfRecords() {
             return numberOfRecords;
         }
